react: simplify — de-fork consent derivation, de-dup policy components, trim surface
Eliminate the deliberately-forked deriveConsentConfig (+ its parity test) by
moving the canonical toPolicyStackConsentConfig into @policystack/core/consent;
@policystack/sdk/consent becomes a thin re-export (public SDK API unchanged).
@policystack/react now imports the one canonical copy, preserving the
react → core-only dependency edge with zero fork to keep in sync — and no
longer needs @policystack/sdk at all (dead devDependency removed).

React-internal trims (no behaviour change):
- Collapse the 40-line PrivacyPolicy/CookiePolicy near-clones onto a shared
  internal <PolicyDocument> parameterized by the compile fn.
- Drop the redundant public renderNode export (renderDocument stays; matches
  @policystack/vue, which keeps renderNode internal).
- Replace useConsent's hand-maintained result type + 9 per-render closure
  wrappers with a Pick<ConsentStore, ...>-derived type returning the store's
  own methods by reference (stable identities).

Net -77 LOC. Verified: vp run -r build/test (960 tests) /check-types green;
vp check clean.
